[refactor pt 1] Move column sorting logic to its own file
- this file is getting hard to follow, so we're giving it the ol' data grid hooks treatment
