+++ packages/eui/src/components/datagrid/body/header/data_grid_header_cell.test.tsx
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-import React, { ReactNode } from 'react';
+import React from 'react';
 import { fireEvent } from '@testing-library/react';
 import {
   render,
@@ -20,7 +20,6 @@ import { mockFocusContext } from '../../utils/__mocks__/focus_context';
 
 import {
   EuiDataGridHeaderCell,
-  useSortingUtils,
   usePopoverArrowNavigation,
 } from './data_grid_header_cell';
 
@@ -53,160 +52,6 @@ describe('EuiDataGridHeaderCell', () => {
   it('renders', () => {
     const { container } = render(<EuiDataGridHeaderCell {...requiredProps} />);
     expect(container.firstChild).toMatchSnapshot();
-  });
-
-  describe('sorting', () => {
-    const onSort = () => {};
-    const columnId = 'test';
-    const mockSortingArgs = {
-      sorting: undefined,
-      id: columnId,
-      showColumnActions: true,
-    };
-
-    const getRender = (node: ReactNode) => render(<>{node}</>).container;
-
-    describe('if the current column is being sorted', () => {
-      it('renders an ascending sort arrow', () => {
-        const { sortingArrow } = renderHook(() =>
-          useSortingUtils({
-            ...mockSortingArgs,
-            sorting: {
-              onSort,
-              columns: [{ id: columnId, direction: 'asc' }],
-            },
-          })
-        ).result.current;
-
-        expect(
-          getRender(sortingArrow).querySelector('[data-euiicon-type="sortUp"]')
-        ).toBeInTheDocument();
-      });
-
-      it('renders a descending sort arrow', () => {
-        const { sortingArrow } = renderHook(() =>
-          useSortingUtils({
-            ...mockSortingArgs,
-            sorting: {
-              onSort,
-              columns: [{ id: columnId, direction: 'desc' }],
-            },
-          })
-        ).result.current;
-
-        expect(
-          getRender(sortingArrow).querySelector(
-            '[data-euiicon-type="sortDown"]'
-          )
-        ).toBeInTheDocument();
-      });
-
-      describe('when only the current column is being sorted', () => {
-        describe('when the header cell has no actions', () => {
-          it('renders aria-sort but not sortingScreenReaderText', () => {
-            const { ariaSort, sortingScreenReaderText } = renderHook(() =>
-              useSortingUtils({
-                ...mockSortingArgs,
-                sorting: {
-                  onSort,
-                  columns: [{ id: columnId, direction: 'asc' }],
-                },
-                showColumnActions: false,
-              })
-            ).result.current;
-
-            expect(ariaSort).toEqual('ascending');
-            expect(getRender(sortingScreenReaderText)).toHaveTextContent('');
-          });
-        });
-
-        describe('when the header cell has actions', () => {
-          it('renders aria-sort and sortingScreenReaderText', () => {
-            const { ariaSort, sortingScreenReaderText } = renderHook(() =>
-              useSortingUtils({
-                ...mockSortingArgs,
-                sorting: {
-                  onSort,
-                  columns: [{ id: columnId, direction: 'desc' }],
-                },
-                showColumnActions: true,
-              })
-            ).result.current;
-
-            expect(ariaSort).toEqual('descending');
-            expect(getRender(sortingScreenReaderText)).toHaveTextContent(
-              'Sorted descending.'
-            );
-          });
-        });
-      });
-    });
-
-    describe('if the current column is not being sorted', () => {
-      it('does not render an arrow even if other columns are sorted', () => {
-        const { sortingArrow } = renderHook(() =>
-          useSortingUtils({
-            ...mockSortingArgs,
-            sorting: { onSort, columns: [{ id: 'other', direction: 'desc' }] },
-          })
-        ).result.current;
-
-        expect(sortingArrow).toBeNull();
-      });
-
-      it('does not render aria-sort or screen reader sorting text', () => {
-        const { ariaSort, sortingScreenReaderText } = renderHook(() =>
-          useSortingUtils(mockSortingArgs)
-        ).result.current;
-
-        expect(ariaSort).toEqual(undefined);
-        expect(getRender(sortingScreenReaderText)).toHaveTextContent('');
-      });
-    });
-
-    describe('when multiple columns are being sorted', () => {
-      it('does not render aria-sort, but renders sorting screen reader text text with a full list of sorted columns', () => {
-        const { result, rerender } = renderHook(useSortingUtils, {
-          initialProps: {
-            ...mockSortingArgs,
-            id: 'A',
-            sorting: {
-              onSort,
-              columns: [
-                { id: 'A', direction: 'asc' },
-                { id: 'B', direction: 'desc' },
-              ],
-            },
-          },
-        });
-
-        expect(result.current.ariaSort).toEqual(undefined);
-        expect(
-          getRender(result.current.sortingScreenReaderText)
-        ).toHaveTextContent(
-          'Sorted by A, ascending, then sorted by B, descending.'
-        );
-
-        // Branch coverage
-        rerender({
-          ...mockSortingArgs,
-          id: 'B',
-          sorting: {
-            onSort,
-            columns: [
-              { id: 'B', direction: 'desc' },
-              { id: 'C', direction: 'asc' },
-              { id: 'A', direction: 'asc' },
-            ],
-          },
-        });
-        expect(
-          getRender(result.current.sortingScreenReaderText)
-        ).toHaveTextContent(
-          'Sorted by B, descending, then sorted by C, ascending, then sorted by A, ascending.'
-        );
-      });
-    });
   });
 
   describe('resizing', () => {
